[jit] Introduce Opcache\Jit attribute to replace @jit
Instead of @jit in the docblock you can now use the <<Opcache\Jit>>
attribute. In addition this attribute allows you to disable the JIT
for a specific function as well with <<Opcache\Jit(false)>>.

Another behavior change is that the Opcache\Jit attribute has precedence
over any other trigger as well. This means you can allow to force JIT
in profiling triggers or disable JIT for a function regardless of trigger.